refac: creating files for screens
