+++ lib/invoices/screens/InvoiceDataScreen.dart
@@ -1,0 +1,32 @@
+import 'package:finance_controlinator_mobile/invoices/domain/sync/InvoiceSync.dart';
+
+import 'package:flutter/material.dart';
+import '../domain/sync/InvoiceSync.dart';
+import 'InvoiceItemsComponent.dart';
+import 'InvoiceOverview.dart';
+
+import 'package:flutter/cupertino.dart';
+
+class InvoiceDataScreen extends StatelessWidget {
+  InvoiceMonthData invoiceMonthData;
+
+  InvoiceDataScreen(this.invoiceMonthData, {Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      children: [
+        Expanded(
+          flex: 40,
+          child: InvoiceOverviewScreen(invoiceMonthData.overview),
+        ),
+        Expanded(
+            flex: 60,
+            child: Padding(
+              padding: const EdgeInsets.only(top: 16),
+              child: InvoiceItemsComponent(invoiceMonthData.invoice.items!),
+            ))
+      ],
+    );
+  }
+}
